Refactor edge rendering code into dedicated renderer
