+++ src/grapheditor.ts
@@ -21,20 +21,17 @@ import { curveBasis } from 'd3-shape';
 import { D3ZoomEvent, zoom, ZoomBehavior, zoomIdentity, zoomTransform, ZoomTransform } from 'd3-zoom';
 import { NodeRenderer } from './rendering/node-renderer';
 import { NodeDropZone } from './drop-zone';
-import { DefaultTextComponentTemplate, DynamicMarkerTemplate, DynamicTextComponentTemplate } from './dynamic-templates/dynamic-template';
+import { DefaultTextComponentTemplate } from './dynamic-templates/dynamic-template';
 import { EdgePathGenerator, SmoothedEdgePathGenerator } from './dynamic-templates/edge-path-generators';
-import { DraggedEdge, Edge, EdgeDragHandle, edgeId, normalizePositionOnLine, PathPositionRotationAndScale, Point, setDefaultEdgeDragHandles, TextComponent } from './edge';
+import { DraggedEdge, Edge, edgeId, Point } from './edge';
 import { GroupingManager } from './grouping';
 import { LinkHandle } from './link-handle';
-import { applyUserLinkHandleCalculationCallback, calculateNearestHandles, getNodeLinkHandles } from './link-handle-helper';
-import { LineAttachementInfo, Marker } from './marker';
 import { Node, NodeMovementInformation } from './node';
 import { GraphObjectCache } from './object-cache';
 import { ExtrasRenderer } from './rendering/extras-renderer';
-import { calculateAngle, calculateRotationTransformationAngle, normalizeVector, RotationVector } from './rotation-vector';
 import { DynymicTemplateRegistry, EdgePathGeneratorRegistry, StaticTemplateRegistry } from './templating';
-import { wrapText } from './textwrap';
-import { Rect, recursiveAttributeGet, squaredPointDistance } from './util';
+import { Rect, squaredPointDistance } from './util';
+import { EdgeRenderer } from './rendering/edge-renderer';
 
 
 const SHADOW_DOM_TEMPLATE = `
@@ -120,7 +117,7 @@ export default class GraphEditor extends HTMLElement {
      * templates get updated!
      */
     public edgePathGeneratorRegistry: EdgePathGeneratorRegistry;
-    private defaultEdgePathGenerator: EdgePathGenerator;
+    public readonly defaultEdgePathGenerator: EdgePathGenerator;
 
     /* Renderers that implement the actual rendering methods. */
 
@@ -141,6 +138,14 @@ export default class GraphEditor extends HTMLElement {
     public nodeRenderer: NodeRenderer;
 
     /**
+     * Renderer for edge specific rendering functionality.
+     *
+     * Do not call methods from this object directly!
+     * Do not replace this object!
+     */
+    public edgeRenderer: EdgeRenderer;
+
+    /**
      * Object responsible for managing node groups and group behaviours.
      *
      * Do not replace this object!
@@ -150,7 +155,7 @@ export default class GraphEditor extends HTMLElement {
     /**
      * The object cache responsible for fast access of nodes and edges.
      */
-    public objectCache: GraphObjectCache; // FIXME make this private again after refactoring works
+    private objectCache: GraphObjectCache;
 
 
     /** Private property to determine if the graph can be drawn. */
@@ -350,6 +355,10 @@ export default class GraphEditor extends HTMLElement {
         return this._edges;
     }
 
+    get draggedEdgeList(): DraggedEdge[] {
+        return this.draggedEdges;
+    }
+
     /**
      * The list of edges.
      *
@@ -442,9 +451,10 @@ export default class GraphEditor extends HTMLElement {
         this.defaultEdgePathGenerator = new SmoothedEdgePathGenerator(curveBasis, true, 10);
         this.edgePathGeneratorRegistry.addEdgePathGenerator('default', this.defaultEdgePathGenerator);
 
-        // TODO make sure to dispose these circular references correctly
+        // FIXME (use weak references) make sure to dispose these circular references correctly
         this.extrasRenderer = new ExtrasRenderer(this);
         this.nodeRenderer = new NodeRenderer(this, this.objectCache);
+        this.edgeRenderer = new EdgeRenderer(this, this.objectCache);
         this.groupingManager = new GroupingManager(this);
 
         this.root = this.attachShadow({ mode: 'open' });
@@ -839,15 +849,22 @@ export default class GraphEditor extends HTMLElement {
      *
      * @param edge edge to add
      * @param redraw if graph should be redrawn (default: `false`)
+     * @param eventSource specify the event source type for the fired events
+     * @returns true if the graph needs to be re-rendered to display changes
      */
-    public addEdge(edge: Edge, redraw: boolean = false): void {
-        this._edges.push(edge);
-        this.objectCache.addEdgeToCache(edge);
-        this.onEdgeCreate(edge, EventSource.API, false);
-        if (redraw) {
-            this.completeRender(false, EventSource.API);
-            this.zoomToBoundingBox(false);
+    public addEdge(edge: Edge, redraw: boolean = false, eventSource: EventSource=EventSource.API): boolean {
+        if (this.onEdgeCreate(edge, eventSource, eventSource !== EventSource.API)) {
+            // event was not cancelled
+            this._edges.push(edge);
+            this.objectCache.addEdgeToCache(edge);
+            if (redraw) {
+                this.completeRender(false, eventSource);
+                this.zoomToBoundingBox(false);
+            } else {
+                return true;
+            }
         }
+        return false;
     }
 
     /**
@@ -865,8 +882,10 @@ export default class GraphEditor extends HTMLElement {
      *
      * @param edge edge to remove
      * @param redraw if the graph should be redrawn (default: `false`)
+     * @param eventSource specify the event source type for the fired events
+     * @returns true if the graph needs to be re-rendered to display changes
      */
-    public removeEdge(edge: Edge | number | string, redraw: boolean = false): void {
+    public removeEdge(edge: Edge | number | string, redraw: boolean = false, eventSource: EventSource=EventSource.API): boolean {
         let edgeIdToDelete: string;
         if (typeof (edge) === 'number') {
             edgeIdToDelete = edge.toString();
@@ -878,14 +897,18 @@ export default class GraphEditor extends HTMLElement {
         const index = this._edges.findIndex((e) => edgeId(e) === edgeIdToDelete);
         if (index >= 0) {
             const removedEdge = this._edges[index];
-            this.onEdgeRemove(removedEdge, EventSource.API, false);
-            this._edges.splice(index, 1);
-            this.objectCache.removeEdgeFromCache(removedEdge);
-            if (redraw) {
-                this.completeRender(false, EventSource.API);
-                this.zoomToBoundingBox(false);
+            if (this.onEdgeRemove(removedEdge, eventSource, eventSource !== EventSource.API)) {
+                this._edges.splice(index, 1);
+                this.objectCache.removeEdgeFromCache(removedEdge);
+                if (redraw) {
+                    this.completeRender(false, eventSource);
+                    this.zoomToBoundingBox(false);
+                } else {
+                    return true;
+                }
             }
         }
+        return false;
     }
 
     /**
@@ -904,6 +927,16 @@ export default class GraphEditor extends HTMLElement {
      */
     public getEdgesByTarget(targetNodeId: number | string): Set<Edge> {
         return this.objectCache.getEdgesByTarget(targetNodeId);
+    }
+
+    /**
+     * Check if an edge is currently dragged by the user.
+     *
+     * @param edgeId the id of the edge to check
+     * @returns true if a dragged edge exists that was created from this edge
+     */
+    public isEdgeCurrentlyDragged(edgeId: string|number): boolean {
+        return this.draggedEdges.some((edge) => edge.createdFrom === edgeId);
     }
 
     /**
@@ -1165,6 +1198,7 @@ export default class GraphEditor extends HTMLElement {
      * @param box a box in graph coordinates
      */
     public zoomToBox(box: Rect): void {
+        // FIXME allow zooming to exact box (without padding applied!)
         const scale = 0.9 * Math.min(this.contentMaxWidth / box.width, this.contentMaxHeight / box.height);
 
         const xCorrection = (-box.x * scale) + ((this.contentMaxWidth - (box.width * scale)) / 2);
@@ -1390,28 +1424,8 @@ export default class GraphEditor extends HTMLElement {
         this.nodeRenderer.completeNodeGroupsRender(this.nodesGroup, this._nodes, forceUpdateTemplates);
 
         // update edges ////////////////////////////////////////////////////////
-        if (forceUpdateTemplates) {
-            this.edgesGroup.selectAll('g.edge-group:not(.dragged)').remove();
-        }
-        const self = this;
-        this.edgesGroup
-            .selectAll<SVGGElement, Edge>('g.edge-group:not(.dragged)')
-            .data<Edge>(this._edges, edgeId)
-            .join(
-                enter => enter.append('g')
-                    .attr('id', (d) => `edge-${edgeId(d)}`)
-                    .classed('edge-group', true)
-                    .each(function (d) {
-                        const edgeGroup = select(this);
-                        edgeGroup.append('path')
-                            .classed('edge', true)
-                            .attr('fill', 'none');
-                    })
-            )
-            .call(self.updateEdgeGroups.bind(this))
-            .call(self.updateEdgePositions.bind(this))
-            .order()
-            .on('click', (event, d) => { this.onEdgeClick.bind(this)(event, d); });
+        this.edgeRenderer.completeEdgeGroupsRender(this.edgesGroup, this._edges, forceUpdateTemplates);
+
 
         this.classesToRemove.clear();
 
@@ -1443,7 +1457,7 @@ export default class GraphEditor extends HTMLElement {
 
             while (!target.empty()) {
                 if (target.classed('node')) {
-                    const id = target.attr('id').replace(/^node-/, '');
+                    const id = target.attr('id').replace(/^node-/, '');  // FIXME use data attr for node ids!
                     const node = this.objectCache.getNode(id);
                     if (node != null && !foundNodes.has(id)) {
                         foundNodes.add(id);
@@ -1567,19 +1581,19 @@ export default class GraphEditor extends HTMLElement {
      * @param force force text rewrap even when text has not changed
      *      (useful if node classes can change text attributes like size)
      */
-    public updateTextElements(force: boolean = false): void {
+    public updateTextElements(force: boolean = false): void { // TODO profile this
         const self = this;
 
         this.nodesGroup
             .selectAll<SVGGElement, Node>('g.node')
             .data<Node>(this._nodes, (d: Node) => d.id.toString())
-            .call(this.updateNodeText.bind(this), force);
+            .call(this.nodeRenderer.updateNodeText.bind(this.nodeRenderer), force);
 
         this.edgesGroup
             .selectAll<SVGGElement, Edge>('g.edge-group:not(.dragged)')
             .data<Edge>(this._edges, edgeId)
             .each(function (d) {
-                self.updateEdgeText(select(this), d, force);
+                self.edgeRenderer.updateEdgeText(select(this), d, force);
             });
         this.onRender(EventSource.API, 'text');
     }
@@ -1645,10 +1659,10 @@ export default class GraphEditor extends HTMLElement {
      *
      * @param nodeId the id of the node to select
      */
-    private getSingleNodeSelection(nodeId: string | number): Selection<SVGGElement, Node, any, unknown> {
+    public getSingleNodeSelection(nodeId: string | number): Selection<SVGGElement, Node, any, unknown> {
         const node = this.objectCache.getNode(nodeId);
         if (node != null) {
-            return this.nodesGroup.select<SVGGElement>(`g.node#node-${nodeId}`).datum(node);
+            return this.nodesGroup.select<SVGGElement>(`g.node#node-${nodeId}`).datum(node); // FIXME use data- attr for node id
         }
         return null;
     }
@@ -1679,52 +1693,6 @@ export default class GraphEditor extends HTMLElement {
         }
     }
 
-
-
-    /**
-     * Update edge groups.
-     *
-     * @param edgeGroupSelection d3 selection of edgeGroups
-     */
-    private updateEdgeGroups(edgeGroupSelection: Selection<SVGGElement, Edge, any, unknown>) {
-        if (edgeGroupSelection == null) {
-            edgeGroupSelection = this.edgesGroup
-                .selectAll<SVGGElement, Edge>('g.edge-group:not(.dragged)')
-                .data<Edge>(this._edges, edgeId);
-        }
-        const self = this;
-        edgeGroupSelection
-            .each(function (d) {
-                self.updateEdgeGroup(select(this), d);
-            })
-            .call(this.updateEdgeGroupClasses.bind(this))
-            .call(this.updateEdgeHighligts.bind(this));
-    }
-
-
-    /**
-     * Update draggededge groups.
-     */
-    public updateDraggedEdgeGroups() { // FIXME visibility
-        this.edgesGroup
-            .selectAll<SVGGElement, DraggedEdge>('g.edge-group.dragged')
-            .data<DraggedEdge>(this.draggedEdges, edgeId)
-            .join(
-                enter => enter.append('g')
-                    .attr('id', (d) => `edge-${edgeId(d)}`)
-                    .classed('edge-group', true)
-                    .classed('dragged', true)
-                    .each(function () {
-                        select(this).append('path')
-                            .classed('edge', true)
-                            .attr('fill', 'none');
-                    })
-            )
-            .call(this.updateEdgeGroupClasses.bind(this))
-            .call(this.updateEdgeGroups.bind(this))
-            .call(this.updateEdgePositions.bind(this));
-    }
-
     /**
      * Update classes of edgeGroups
      *
@@ -1735,681 +1703,10 @@ export default class GraphEditor extends HTMLElement {
         if (edgeGroupSelection == null) {
             edgeGroupSelection = this.getEdgeSelection();
         }
-        edgeGroupSelection.classed('ghost', (d) => {
-            const id = edgeId(d);
-            return this.draggedEdges.some((edge) => edge.createdFrom === id);
-        });
-        if (this.classesToRemove != null) {
-            this.classesToRemove.forEach((className) => {
-                edgeGroupSelection.classed(className, () => false);
-            });
-        }
-        if (this.classes != null) {
-            this.classes.forEach((className) => {
-                edgeGroupSelection.classed(className, (d) => {
-                    if (this.setEdgeClass != null) {
-                        return this.setEdgeClass(className, d, this.objectCache.getNode(d.source),
-                            (d.target != null) ? this.objectCache.getNode(d.target) : null);
-                    }
-                    return false;
-                });
-            });
-        }
+        this.edgeRenderer.updateEdgeGroupClasses(edgeGroupSelection, this.classesToRemove);
         if (calledDirectly) {
             this.onRender(EventSource.API, 'classes');
         }
-    }
-
-    /**
-     * Update edge path and marker positions.
-     *
-     * @param edgeGroupSelection d3 selection
-     */
-    private updateEdgePositions(edgeGroupSelection: Selection<SVGGElement, Edge, any, unknown>) {
-        const self = this;
-        edgeGroupSelection.select<SVGPathElement>('path.edge')
-            .call(this.updateEdgeLinkHandles.bind(this))
-            .call(this.updateEdgePath.bind(this));
-        edgeGroupSelection.each(function (d) {
-            self.updateEdgeTextPositions(select(this), d);
-        });
-        edgeGroupSelection.each(function (d) {
-            select(this).selectAll('g.marker:not(.marker-special)').data(d.markers != null ? d.markers : [])
-                .call(self.updateMarkerPositions.bind(self));
-        }).each(function (d) {
-            self.updateEndMarkerPositions(select(this), d);
-        }).each(function (d) {
-            // update link handle position
-            setDefaultEdgeDragHandles(d);
-            select(this).selectAll('g.link-handle').data(d.dragHandles)
-                .call(self.updateMarkerPositions.bind(self));
-        });
-    }
-
-    /**
-     * Update markers and path attributes.
-     *
-     * @param edgeGroupSelection d3 selection of single edge group
-     * @param d edge datum
-     */
-    private updateEdgeGroup(edgeGroupSelection: Selection<SVGGElement, Edge, any, unknown>, d: Edge) {
-        const pathSelection = edgeGroupSelection.select<SVGPathElement>('path.edge:not(.dragged)').datum(d);
-        pathSelection.attr('stroke', 'black');
-
-        // update markers
-        this.updateEndMarkers(edgeGroupSelection, d);
-        edgeGroupSelection.selectAll<SVGGElement, Marker>('g.marker:not(.marker-special)')
-            .data(d.markers != null ? d.markers : [])
-            .join(
-                enter => enter.append('g')
-                    .classed('marker', true)
-            )
-            .call(this.updateMarker.bind(this), d);
-
-        // update edge drag handles
-        setDefaultEdgeDragHandles(d);
-        const edgeDragHandles = edgeGroupSelection.selectAll<SVGGElement, EdgeDragHandle>('g.link-handle')
-            .data(d.dragHandles)
-            .join(
-                enter => enter.append('g')
-                    .classed('link-handle', true)
-            )
-            .call(this.updateMarker.bind(this), d)
-            .raise(); // raise the drag handles to the top of the edge
-
-        this.updateEdgeText(edgeGroupSelection, d);
-        this.extrasRenderer.updateDynamicProperties(edgeGroupSelection);
-
-        edgeDragHandles.call(
-            drag<SVGGElement, EdgeDragHandle, { edge: DraggedEdge; capturingGroup?: string; isReversedEdge: boolean }>()
-                .subject((e, h) => {
-                    if (this._edgeDragInteraction === 'none') {
-                        return; // edge dragging is disabled
-                    }
-                    const event = e as unknown as Event;
-                    const handle = h as unknown as EdgeDragHandle;
-                    const edge = d;
-                    let sourceNode: Node;
-                    if ((handle).isReverseHandle ?? false) {
-                        // a reverse handle flips the edge direction
-                        sourceNode = this.getNode(edge.target);
-                    } else {
-                        sourceNode = this.getNode(edge.source);
-                    }
-                    const groupCapturingEdge = this.groupingManager.getGroupCapturingOutgoingEdge(sourceNode);
-                    if (groupCapturingEdge != null && groupCapturingEdge !== sourceNode.id.toString()) {
-                        const groupNode = this.getNode(groupCapturingEdge);
-                        if (groupNode != null) {
-                            const newEdge = this.createDraggedEdgeFromExistingEdge(event, edge);
-                            newEdge.source = groupCapturingEdge;
-                            return {
-                                edge: newEdge,
-                                capturingGroup: groupCapturingEdge,
-                                isReversedEdge: handle.isReverseHandle ?? false,
-                            };
-                        }
-                    }
-                    return {
-                        edge: this.createDraggedEdgeFromExistingEdge(event, edge, handle.isReverseHandle ?? false),
-                        capturingGroup: sourceNode.id.toString(),
-                        isReversedEdge: handle.isReverseHandle ?? false,
-                    };
-                })
-                .container(() => this.edgesGroup.node() as any)
-                .on('start', () => this.completeRender(false, EventSource.USER_INTERACTION))
-                .on('drag', (event) => {
-                    this.updateDraggedEdge(event as any, (event as any).subject.edge, (event as any).subject.capturingGroup);
-                    this.updateDraggedEdgeGroups();
-                })
-                .on('end', (event) => {
-                    this.dropDraggedEdge(event as any, (event as any).subject.edge, (event as any).subject.isReversedEdge);
-                })
-        );
-    }
-
-    /**
-     * Update all edge texts in a edge group.
-     *
-     * @param edgeGroupSelection d3 selection of single edge group
-     * @param d edge datum
-     * @param force force text to re-wrap
-     */
-    private updateEdgeText(edgeGroupSelection: Selection<SVGGElement, Edge, any, unknown>, d: Edge, force: boolean = false) {
-        const self = this;
-        edgeGroupSelection.each(function (edge) {
-            const textSelection = select(this).selectAll<SVGGElement, TextComponent>('g.text-component')
-                .data(edge.texts != null ? edge.texts : [])
-                .join(enter => enter.append('g').classed('text-component', true))
-                .each(function (textComponent) {
-                    const g: Selection<SVGGElement, TextComponent, any, unknown> = select(this).datum<TextComponent>(textComponent);
-                    const templateId = textComponent.template ?? 'default-textcomponent';
-                    self.extrasRenderer.updateContentTemplate<TextComponent>(g, templateId, 'textcomponent', true, edge);
-                    const dynTemplate = self.dynamicTemplateRegistry.getDynamicTemplate<DynamicTextComponentTemplate>(templateId);
-                    try {
-                        dynTemplate?.updateTemplate(g, self, { parent: edge });
-                    } catch (error) {
-                        console.error(`An error occured updating the text component in edge ${edgeId(edge)} before text wrapping`, textComponent, error);
-                    }
-                })
-                .attr('data-click', (t) => t.clickEventKey);
-
-            textSelection.select('text')
-                .classed('text', true)
-                .attr('width', (t) => t.width)
-                .attr('height', (t) => t.height)
-                .each(function (text) {
-                    let newText = '';
-                    if (text.value != null) {
-                        newText = text.value;
-                    } else {
-                        newText = recursiveAttributeGet(d, text.attributePath)?.toString();
-                    }
-                    if (newText == null) {
-                        newText = '';
-                    }
-                    // make sure it is a string
-                    newText = newText.toString();
-                    wrapText(this as SVGTextElement, newText, force);
-                });
-            textSelection.each(function (textComponent) {
-                const g: Selection<SVGGElement, TextComponent, any, unknown> = select(this).datum<TextComponent>(textComponent);
-                const templateId = textComponent.template ?? 'default-textcomponent';
-                const dynTemplate = self.dynamicTemplateRegistry.getDynamicTemplate<DynamicTextComponentTemplate>(templateId);
-                try {
-                    dynTemplate?.updateAfterTextwrapping(g, self, { parent: edge });
-                } catch (error) {
-                    console.error(`An error occured updating the text component in edge ${edgeId(edge)} after text wrapping`, textComponent, error);
-                }
-            });
-            const path = edgeGroupSelection.select<SVGPathElement>('path.edge');
-            textSelection.call(drag()
-                .subject((event, text) => {
-                    if (!((text as unknown as TextComponent).draggable ?? true)) {
-                        return; // text component is not draggable
-                    }
-                    return text;
-                })
-                .on('start', (event, text) => {
-                    self.onEdgeTextDrag('start', text as unknown as TextComponent, edge, EventSource.USER_INTERACTION);
-                })
-                .on('drag', ((event, text: TextComponent) => {
-                    const length = path.node().getTotalLength();
-                    const positionOnLine = normalizePositionOnLine(text.positionOnLine);
-                    const absolutePositionOnLine = self.calculateAbsolutePositionOnLine(length, positionOnLine, text.absolutePositionOnLine);
-                    const referencePoint = path.node().getPointAtLength(absolutePositionOnLine);
-                    text.offsetX = event.x - referencePoint.x;
-                    text.offsetY = event.y - referencePoint.y;
-                    self.onEdgeTextPositionChange(text, edge);
-                    self.updateEdgeTextPositions(edgeGroupSelection, edge);
-                }) as any) // FIXME: remove type hack when types are up to date
-                .on('end', ((event, text: TextComponent) => {
-                    self.onEdgeTextDrag('end', text, edge, EventSource.USER_INTERACTION);
-                }) as any) // FIXME: remove type hack when types are up to date
-            );
-        });
-    }
-
-    /**
-     * Calculate the attachement vector for a marker.
-     *
-     * @param startingAngle the line angle for the marker
-     * @param marker the selection of a single marker
-     * @param strokeWidth the current stroke width
-     */
-    private calculateLineAttachementVector(startingAngle: number | RotationVector, markerSelection: Selection<SVGGElement, Marker, any, unknown>, strokeWidth: number) {
-        if (markerSelection.empty()) {
-            return { dx: 0, dy: 0 };
-        }
-        const marker = markerSelection.datum();
-        let attachementPointInfo: LineAttachementInfo;
-        if (marker.isDynamicTemplate) {
-            const dynTemplate = this.dynamicTemplateRegistry.getDynamicTemplate<DynamicMarkerTemplate>(marker.template);
-            try {
-                attachementPointInfo = dynTemplate?.getLineAttachementInfo(markerSelection);
-            } catch (error) {
-                console.error('An error occured while calculating the line attachement info for an edge marker!', marker, error);
-            }
-        } else {
-            attachementPointInfo = this.staticTemplateRegistry.getMarkerAttachementPointInfo(marker.template);
-        }
-        if (attachementPointInfo != null) {
-            let scale = 1;
-            if (marker.scale != null) {
-                scale *= marker.scale;
-            }
-            if (Boolean(marker.scaleRelative)) {
-                scale *= strokeWidth;
-            }
-            if (typeof startingAngle === 'number') {
-                return attachementPointInfo.getRotationVector(startingAngle, scale);
-            } else {
-                const normalAngle = calculateAngle(startingAngle);
-                return attachementPointInfo.getRotationVector(normalAngle, scale);
-            }
-        }
-        return { dx: 0, dy: 0 };
-    }
-
-    /**
-     * Calculate the link handles for each edge and store them into the edge.
-     *
-     * @param edgeSelection d3 selection of edges to update with bound data
-     */
-    private updateEdgeLinkHandles(edgeSelection: Selection<SVGPathElement, Edge | DraggedEdge, any, unknown>) {
-        edgeSelection.each(edge => {
-
-            const sourceNodeSelection = this.getSingleNodeSelection(edge.source);
-            const targetNodeSelection = (edge.target != null) ? this.getSingleNodeSelection(edge.target) : null;
-            let initialSourceHandles, initialTargetHandles;
-            try {
-                initialSourceHandles = getNodeLinkHandles(sourceNodeSelection, this.staticTemplateRegistry, this.dynamicTemplateRegistry, this);
-            } catch (error) {
-                console.error(`An error occured while calculating the link handles for node ${edge.source}!`, error);
-            }
-            try {
-                initialTargetHandles = getNodeLinkHandles(targetNodeSelection, this.staticTemplateRegistry, this.dynamicTemplateRegistry, this);
-            } catch (error) {
-                console.error(`An error occured while calculating the link handles for node ${edge.target}!`, error);
-            }
-
-            let sourceNode: Node, targetNode: Node | Point;
-            if (sourceNodeSelection != null && !sourceNodeSelection.empty()) {
-                sourceNode = sourceNodeSelection.datum();
-            } else {
-                console.warn('Attempting to render edge without a valid source!');
-            }
-            if (targetNodeSelection != null && !targetNodeSelection.empty()) {
-                targetNode = targetNodeSelection.datum();
-            } else if (edge.currentTarget != null) {
-                targetNode = edge.currentTarget as Point;
-            } else {
-                console.warn('Attempting to render edge without a valid target!');
-            }
-
-            const newHandles = applyUserLinkHandleCalculationCallback(
-                edge,
-                initialSourceHandles,
-                sourceNode,
-                initialTargetHandles,
-                targetNode,
-                this.calculateLinkHandlesForEdge
-            );
-
-            const nearestHandles = calculateNearestHandles(newHandles.sourceHandles, sourceNode, newHandles.targetHandles, targetNode);
-            edge.sourceHandle = nearestHandles.sourceHandle;
-            edge.targetHandle = nearestHandles.targetHandle;
-        });
-    }
-
-    /**
-     * Update existing edge path.
-     *
-     * @param edgeSelection d3 selection of edges to update with bound data
-     */
-    private updateEdgePath(edgeSelection: Selection<SVGPathElement, Edge, any, unknown>) {
-        const self = this;
-        edgeSelection.each(function (edge) {
-            const singleEdgeSelection = select(this).datum(edge);
-            const strokeWidth: number = parseFloat(singleEdgeSelection.style('stroke-width').replace(/px/, ''));
-            // eslint-disable-next-line complexity
-            singleEdgeSelection.attr('d', (d) => {
-                let sourceCoordinates: Point = d.source != null ? self.objectCache.getNode(d.source) : null;
-                let targetCoordinates: Point = d.target != null ? self.objectCache.getNode(d.target) : null;
-
-                if (sourceCoordinates == null) {
-                    sourceCoordinates = { x: 0, y: 0 };
-                }
-                if (targetCoordinates == null) {
-                    if (d.currentTarget != null) {
-                        targetCoordinates = d.currentTarget;
-                    } else {
-                        targetCoordinates = { x: 0, y: 1 };
-                    }
-                }
-
-                // apply link handle offsets
-                if (d.sourceHandle != null) {
-                    sourceCoordinates = {
-                        x: sourceCoordinates.x + d.sourceHandle.x,
-                        y: sourceCoordinates.y + d.sourceHandle.y,
-                    };
-                }
-                if (d.targetHandle != null) {
-                    targetCoordinates = {
-                        x: targetCoordinates.x + d.targetHandle.x,
-                        y: targetCoordinates.y + d.targetHandle.y,
-                    };
-                }
-
-                let sourceHandleNormal: RotationVector;
-                let targetHandleNormal: RotationVector;
-
-                // rotation vector between edge start and end
-                let baseNormal: RotationVector = {
-                    dx: sourceCoordinates.x - targetCoordinates.x,
-                    dy: sourceCoordinates.y - targetCoordinates.y,
-                };
-                if (baseNormal.dx === 0 && baseNormal.dy === 0) {
-                    baseNormal.dx = 1;
-                }
-                baseNormal = normalizeVector(baseNormal);
-
-                if (d.sourceHandle?.normal != null) {
-                    sourceHandleNormal = d.sourceHandle.normal;
-                } else {
-                    sourceHandleNormal = baseNormal;
-                }
-                if (d.targetHandle?.normal != null) {
-                    targetHandleNormal = d.targetHandle.normal;
-                } else {
-                    targetHandleNormal = { dx: -baseNormal.dx, dy: -baseNormal.dy };
-                }
-
-                // calculate path
-                const points: { x: number; y: number;[prop: string]: any }[] = [];
-
-                // Calculate line attachement point for startMarker
-                let startAttachementPointVector: RotationVector = { dx: 0, dy: 0 };
-                if (d.markerStart != null) {
-                    const markerSelection: Selection<SVGGElement, Marker, any, unknown> = select(this.parentNode as any)
-                        .select<SVGGElement>('g.marker.marker-start')
-                        .datum(d.markerStart);
-                    startAttachementPointVector = self.calculateLineAttachementVector(sourceHandleNormal, markerSelection, strokeWidth);
-                }
-
-                points.push({
-                    x: sourceCoordinates.x - startAttachementPointVector.dx,
-                    y: sourceCoordinates.y - startAttachementPointVector.dy,
-                });
-                points.push({
-                    x: sourceCoordinates.x - startAttachementPointVector.dx + (sourceHandleNormal.dx * 10),
-                    y: sourceCoordinates.y - startAttachementPointVector.dy + (sourceHandleNormal.dy * 10),
-                });
-
-                // Calculate line attachement point for endMarker
-                let endAttachementPointVector: RotationVector = { dx: 0, dy: 0 };
-                if (d.markerEnd != null) {
-                    const markerSelection: Selection<SVGGElement, Marker, any, unknown> = select(this.parentNode as any)
-                        .select<SVGGElement>('g.marker.marker-end')
-                        .datum(d.markerEnd);
-                    endAttachementPointVector = self.calculateLineAttachementVector(targetHandleNormal, markerSelection, strokeWidth);
-                }
-
-                if (d.target != null) {
-                    points.push({
-                        x: targetCoordinates.x - endAttachementPointVector.dx + (targetHandleNormal.dx * 10),
-                        y: targetCoordinates.y - endAttachementPointVector.dy + (targetHandleNormal.dy * 10),
-                    });
-                    points.push({
-                        x: targetCoordinates.x - endAttachementPointVector.dx,
-                        y: targetCoordinates.y - endAttachementPointVector.dy,
-                    });
-                } else {
-                    points.push(targetCoordinates);
-                }
-                const pathGenerator = self.edgePathGeneratorRegistry.getEdgePathGenerator(d.pathType) ?? self.defaultEdgePathGenerator;
-                let path: string;
-                try {
-                    path = pathGenerator.generateEdgePath(points[0], points[points.length - 1], sourceHandleNormal, (d.target != null) ? targetHandleNormal : null);
-                } catch (error) {
-                    console.error(`An error occurred while generating the edge path for the edge ${edgeId(edge)}`, error);
-                    path = self.defaultEdgePathGenerator.generateEdgePath(points[0], points[points.length - 1], sourceHandleNormal, (d.target != null) ? targetHandleNormal : null);
-                }
-                return path;
-            });
-        });
-    }
-
-    /**
-     * Update existing edge marker.
-     *
-     * @param markerSelection d3 selection
-     * @param edge the edge datum this marker belongs to
-     */
-    private updateMarker(markerSelection: Selection<SVGGElement, Marker, any, unknown>, edge: Edge) {
-        const self = this;
-        markerSelection
-            .attr('data-click', (d) => d.clickEventKey)
-            .each(function (marker) {
-                const templateId = self.staticTemplateRegistry.getMarkerTemplateId(marker.template);
-                self.extrasRenderer.updateContentTemplate<Marker>(select(this), templateId, 'marker');
-                if (marker.isDynamicTemplate) {
-                    const g = select(this).datum(marker);
-                    const dynTemplate = self.dynamicTemplateRegistry.getDynamicTemplate<DynamicMarkerTemplate>(templateId);
-                    if (dynTemplate != null) {
-                        try {
-                            dynTemplate.updateTemplate(g, self, { parent: edge });
-                        } catch (error) {
-                            console.error(`An error occured while updating the dynamic marker template in edge ${edgeId(edge)}!`, error);
-                        }
-                    }
-                }
-            });
-    }
-
-
-    /**
-     * Update edge-end and edge-start marker.
-     *
-     * @param edgeGroupSelection d3 selection of single edge group
-     * @param d edge datum
-     */
-    private updateEndMarkers(edgeGroupSelection: Selection<SVGGElement, Edge, any, unknown>, d: Edge) {
-        this.updateEndMarker(edgeGroupSelection, d.markerStart, 'marker-start', d);
-        this.updateEndMarker(edgeGroupSelection, d.markerEnd, 'marker-end', d);
-    }
-
-    /**
-     * Update a specific edge end marker (either start or end marker).
-     *
-     * @param edgeGroupSelection d3 selection of single edge group
-     * @param marker the special end marker
-     * @param markerClass the css class to select for
-     * @param edge the edge datum this marker belongs to
-     */
-    private updateEndMarker(edgeGroupSelection: Selection<SVGGElement, Edge, any, unknown>, marker: Marker, markerClass: string, edge: Edge) {
-        if (marker == null) {
-            // delete
-            edgeGroupSelection.select('g.marker.marker-end').remove();
-            return;
-        }
-        let markerEndSelection: Selection<SVGGElement, Marker, any, unknown>;
-        markerEndSelection = edgeGroupSelection.select<SVGGElement>(`g.marker.${markerClass}`)
-            .datum<Marker>(marker);
-        if (markerEndSelection.empty()) {
-            // create
-            markerEndSelection = edgeGroupSelection.append('g')
-                .classed('marker', true)
-                .classed(markerClass, true)
-                .classed('marker-special', true)
-                .datum<Marker>(marker);
-        }
-        this.updateMarker(markerEndSelection, edge);
-    }
-
-    /**
-     * Calculate the safe absolutePositionOnLine value for the given path length.
-     *
-     * If absolutePositidragHandlesonOnLine is negative it is counted from the end of the path.
-     * If absolutePositidragHandlesonOnLine exceeds the path length positionOnLine will be used as fallback.
-     *
-     * @param pathLength the length of the path
-     * @param positionOnLine the relative position on the line (between 0 and 1)
-     * @param absolutePositidragHandlesonOnLine the absolute position on line (between 0 and length)
-     * @returns the positive absolute positionOnLine to be used with `path.getPointAtLength(absolutePositionOnLine)`.
-     */
-    public calculateAbsolutePositionOnLine(pathLength: number, positionOnLine: number, absolutePositionOnLine?: number): number {
-        let result = null;
-        if (absolutePositionOnLine != null) {
-            if (absolutePositionOnLine < 0) {
-                // actually a substraction...
-                result = pathLength + absolutePositionOnLine;
-            } else {
-                result = absolutePositionOnLine;
-            }
-        }
-
-        // else case & sanity checks for if case
-        if (result == null || result < 0 || result > pathLength) {
-            // always fall back to relative position
-            result = pathLength * positionOnLine;
-        }
-        return result;
-    }
-
-    /**
-     * Calculate a normal vector pointing in the direction of the path at the positonOnLine.
-     *
-     * @param path the path object
-     * @param absolutePositionOnLine the absolute position on the path (between 0 and length)
-     * @param point the point at positionOnLine (will be calculated if not supplied)
-     * @param length the length of the path (will be calculated if not supplied)
-     */
-    public calculatePathNormalAtPosition(path: SVGPathElement, absolutePositionOnLine: number, point?: DOMPoint, length?: number): RotationVector {
-        if (length == null) {
-            length = path.getTotalLength();
-        }
-        if (point == null) {
-            point = path.getPointAtLength(absolutePositionOnLine);
-        }
-        const isSecondHalve = absolutePositionOnLine > (length / 2);
-        // calculate a second point at a small offset
-        const epsilon = Math.min(1, length / 100);
-        const delta = isSecondHalve ? -epsilon : epsilon;
-        const point2 = path.getPointAtLength(absolutePositionOnLine + delta);
-        return normalizeVector({
-            dx: isSecondHalve ? (point.x - point2.x) : (point2.x - point.x),
-            dy: isSecondHalve ? (point.y - point2.y) : (point2.y - point.y),
-        });
-    }
-
-    /**
-     * Calculate the transformation attribute for a path object placed on an edge.
-     *
-     * @param point the path object position position
-     * @param pathObject the path object to place
-     * @param strokeWidth the stroke width of the edge
-     * @param normal the normal vector of the edge at the path object position
-     */
-    private calculatePathObjectTransformation(point: { x: number; y: number }, pathObject: PathPositionRotationAndScale, strokeWidth: number, normal: RotationVector) {
-        let transform = `translate(${point.x},${point.y})`;
-        if (pathObject.scale != null) {
-            let scale = pathObject.scale;
-            if (Boolean(pathObject.scaleRelative)) {
-                scale *= strokeWidth;
-            }
-            if (scale !== 1) {
-                transform += `scale(${scale})`;
-            }
-        }
-        const angle = calculateRotationTransformationAngle(pathObject, normal);
-        if (angle !== 0) {
-            transform += `rotate(${angle})`;
-        }
-        return transform;
-    }
-
-    /**
-     * Update positions of edge-end and edge-start marker.
-     *
-     * @param edgeGroupSelection d3 selection of single edge group
-     * @param d edge datum
-     */
-    private updateEndMarkerPositions(edgeGroupSelection: Selection<SVGGElement, Edge, any, unknown>, d: Edge) {
-
-        // calculate position size and rotation
-        const path = edgeGroupSelection.select<SVGPathElement>('path.edge');
-        const length = path.node().getTotalLength();
-        const strokeWidth: number = parseFloat(path.style('stroke-width').replace(/px/, ''));
-
-        if (d.markerStart != null) {
-            this.updateEndMarkerPosition(path, length, 0, d.markerStart, d.sourceHandle, 'marker-start', strokeWidth, edgeGroupSelection);
-        }
-
-        if (d.markerEnd != null) {
-            this.updateEndMarkerPosition(path, length, 1, d.markerEnd, d.targetHandle, 'marker-end', strokeWidth, edgeGroupSelection);
-        }
-    }
-
-    /**
-     * Update a single end marker position (either start or end marker).
-     *
-     * @param path the path selection
-     * @param length the path length
-     * @param positionOnLine positionOnLine at the marker
-     * @param marker the marker
-     * @param handle the link handle at the path end of the marker; can be `null`
-     * @param markerClass the class of the marker
-     * @param strokeWidth the edge stroke width
-     * @param edgeGroupSelection d3 selection of a single edge group
-     */
-    // eslint-disable-next-line complexity
-    private updateEndMarkerPosition(
-        path: Selection<SVGPathElement, Edge, any, unknown>,
-        length: number, positionOnLine: number,
-        marker: Marker, handle: LinkHandle, markerClass: string,
-        strokeWidth: number,
-        edgeGroupSelection: Selection<SVGGElement, Edge, any, unknown>
-    ) {
-        const markerSelection: Selection<SVGGElement, Marker, any, unknown> = edgeGroupSelection
-            .select<SVGGElement>(`g.marker.${markerClass}`)
-            .datum(marker);
-        // path end point
-        const absolutePositionOnLine = this.calculateAbsolutePositionOnLine(length, positionOnLine);
-        const pathPointA = path.node().getPointAtLength(absolutePositionOnLine);
-        // calculate angle for marker
-        let markerStartingNormal: RotationVector;
-        if (handle?.normal?.dx !== 0 || handle?.normal?.dy !== 0) {
-            markerStartingNormal = handle?.normal;
-        }
-        if (markerClass === 'marker-end' && markerStartingNormal != null) {
-            markerStartingNormal = { dx: -markerStartingNormal.dx, dy: -markerStartingNormal.dy };
-        }
-        if (markerStartingNormal == null) {
-            // no link handle for marker present, calculate starting angle from path
-            markerStartingNormal = this.calculatePathNormalAtPosition(path.node(), absolutePositionOnLine, pathPointA, length);
-        }
-        // calculate marker offset
-        const attachementPointVector: RotationVector = this.calculateLineAttachementVector(markerStartingNormal, markerSelection, strokeWidth);
-        const point = {
-            x: pathPointA.x + (positionOnLine === 0 ? attachementPointVector.dx : -attachementPointVector.dx),
-            y: pathPointA.y + (positionOnLine === 0 ? attachementPointVector.dy : -attachementPointVector.dy),
-        };
-        let markerTemplateStartingNormal: RotationVector;
-        // flip normal for markerStart
-        if (markerClass === 'marker-start' && markerStartingNormal != null) {
-            markerTemplateStartingNormal = { dx: -markerStartingNormal.dx, dy: -markerStartingNormal.dy };
-        } else {
-            markerTemplateStartingNormal = markerStartingNormal;
-        }
-        // calculate marker transformation
-        const transformEnd = this.calculatePathObjectTransformation(point, marker, strokeWidth, markerTemplateStartingNormal);
-        // apply transformation
-        markerSelection.attr('transform', transformEnd);
-    }
-
-    /**
-     * Update all edge marker positions
-     *
-     * @param markerSelection d3 selection
-     */
-    private updateMarkerPositions(markerSelection: Selection<SVGGElement, Marker, any, unknown>) {
-        const self = this;
-        markerSelection.each(function (d) {
-            const parent = select(this.parentElement);
-            const marker = select(this);
-            const path = parent.select<SVGPathElement>('path.edge');
-            const length = path.node().getTotalLength();
-            const strokeWidth: number = parseFloat(path.style('stroke-width').replace(/px/, ''));
-            const positionOnLine = normalizePositionOnLine(d.positionOnLine);
-            const absolutePositionOnLine = self.calculateAbsolutePositionOnLine(length, positionOnLine, d.absolutePositionOnLine);
-
-            const point = path.node().getPointAtLength(absolutePositionOnLine);
-            const normal = self.calculatePathNormalAtPosition(path.node(), absolutePositionOnLine, point, length);
-            const transform = self.calculatePathObjectTransformation(point, d, strokeWidth, normal);
-
-            marker.attr('transform', transform);
-        });
     }
 
     /**
@@ -2446,170 +1743,6 @@ export default class GraphEditor extends HTMLElement {
     }
 
     /**
-     * Update all edge text positions in a edge group.
-     *
-     * @param edgeGroupSelection d3 selection of single edge group
-     * @param d edge datum
-     */
-    private updateEdgeTextPositions(edgeGroupSelection: Selection<SVGGElement, Edge, any, unknown>, d: Edge) {
-        const self = this;
-        const path = edgeGroupSelection.select<SVGPathElement>('path.edge');
-        const length = path.node().getTotalLength();
-        const strokeWidth: number = parseFloat(path.style('stroke-width').replace(/px/, ''));
-        const textSelection = edgeGroupSelection.selectAll<SVGGElement, TextComponent>('g.text-component')
-            .data<TextComponent>(d.texts != null ? d.texts : []);
-
-        // calculate the node bounding boxes for collision detection
-        let sourceBB = { x: 0, y: 0, width: 0, height: 0 };
-        let targetBB = { x: 0, y: 0, width: 0, height: 0 };
-        try {
-            const sourceNode = this.objectCache.getNode(d.source);
-            const targetNode = this.objectCache.getNode(d.target);
-            const sourceNodeBB = this.objectCache.getNodeBBox(d.source);
-            const targetNodeBB = this.objectCache.getNodeBBox(d.target);
-            // add node position to bounding box
-            sourceBB = {
-                x: sourceNodeBB.x + sourceNode.x,
-                y: sourceNodeBB.y + sourceNode.y,
-                width: sourceNodeBB.width,
-                height: sourceNodeBB.height,
-            };
-            targetBB = {
-                x: targetNodeBB.x + targetNode.x,
-                y: targetNodeBB.y + targetNode.y,
-                width: targetNodeBB.width,
-                height: targetNodeBB.height,
-            };
-        } catch (error) {
-            // use line endpoints as fallback
-            const sourceEndpoint = path.node().getPointAtLength(0);
-            const targetEndpoint = path.node().getPointAtLength(0);
-            sourceBB.x = sourceEndpoint.x;
-            sourceBB.y = sourceEndpoint.y;
-            targetBB.x = targetEndpoint.x;
-            targetBB.y = targetEndpoint.y;
-        }
-
-        // update text selections
-        // eslint-disable-next-line complexity
-        textSelection.each(function (t) {
-            const text = select(this);
-            const positionOnLine = normalizePositionOnLine(t.positionOnLine);
-            const absolutePositionOnLine = self.calculateAbsolutePositionOnLine(length, positionOnLine, t.absolutePositionOnLine);
-            const pathPoint = path.node().getPointAtLength(absolutePositionOnLine);
-
-            const edgeNormal = self.calculatePathNormalAtPosition(path.node(), absolutePositionOnLine, pathPoint, length);
-
-            // factor in offset coordinates of text component
-            const referencePoint = {
-                x: pathPoint.x + (t.offsetX ?? 0),
-                y: pathPoint.y + (t.offsetY ?? 0),
-            };
-
-            // apply transformation fisrt
-            const initialTransform = self.calculatePathObjectTransformation(referencePoint, t, strokeWidth, edgeNormal);
-            text.attr('transform', initialTransform);
-
-            // calculate center of nearest node (line distance, not euklidean distance)
-            const nodeBB = (absolutePositionOnLine > (length / 2)) ? targetBB : sourceBB;
-            const nodeCenter = {
-                x: nodeBB.x + (nodeBB.width / 2),
-                y: nodeBB.y + (nodeBB.height / 2),
-            };
-
-            // use node center point to calculate the angle with the reference point
-            const normal = {
-                dx: nodeCenter.x - referencePoint.x,
-                dy: nodeCenter.y - referencePoint.y,
-            };
-            let angle = calculateAngle(normal);
-            if (angle < 0) {
-                angle += 380;
-            }
-            if (angle > 360) {
-                angle -= 360;
-            }
-
-            let bbox: Rect = text.node().getBBox();
-
-            if (initialTransform.includes('scale') || initialTransform.includes('rotate')) {
-                const svgNode = text.node();
-                const ctm = (svgNode.parentElement as unknown as SVGGElement).getScreenCTM().inverse().multiply(svgNode.getScreenCTM());
-                bbox = self.transformBBox(bbox, ctm);
-            } else {
-                bbox = {
-                    x: referencePoint.x + bbox.x,
-                    y: referencePoint.y + bbox.y,
-                    width: bbox.width,
-                    height: bbox.height,
-                };
-            }
-
-            const targetPoint: Point = {
-                x: referencePoint.x,
-                y: referencePoint.y,
-            };
-
-            let deltaX = 0;
-            let deltaY = 0;
-            if (angle > 0 && angle < 180) {
-                // bottom of the text (possibly) overlaps
-                let delta = (nodeBB.y) - (bbox.y + bbox.height);
-                if (t.padding) {
-                    delta -= t.padding;
-                }
-                if (delta < 0) {
-                    deltaY = delta;
-                }
-            }
-            if (angle > 180 && angle < 360) {
-                // top of the text (possibly) overlaps
-                let delta = (nodeBB.y + nodeBB.height) - bbox.y;
-                if (t.padding) {
-                    delta += t.padding;
-                }
-                if (delta > 0) {
-                    deltaY = delta;
-                }
-            }
-            if (angle > 90 && angle < 270) {
-                // left side of text (possibly) overlaps
-                let delta = (nodeBB.x + nodeBB.width) - bbox.x;
-                if (t.padding) {
-                    delta += t.padding;
-                }
-                if (delta > 0) { // only update target if text actually overlaps
-                    deltaX = delta;
-                }
-            }
-            if (angle > 270 || angle < 90) {
-                // right side of text (possibly) overlaps
-                let delta = (nodeBB.x) - (bbox.x + bbox.width);
-                if (t.padding) {
-                    delta -= t.padding;
-                }
-                if (delta < 0) { // only update target if text actually overlaps
-                    deltaX = delta;
-                }
-            }
-
-            // only adjust for actual overlap in both directions
-            if (deltaX !== 0 && deltaY !== 0) {
-                // only adjust in one direction
-                if ((angle > 45 && angle < 135) || (angle > 225 && angle < 315)) {
-                    targetPoint.y += deltaY;
-                } else {
-                    targetPoint.x += deltaX;
-                }
-            }
-
-            const finalTransform = initialTransform.replace(/^translate\([^\)]*\)/, `translate(${targetPoint.x},${targetPoint.y})`);
-            text.attr('transform', finalTransform);
-        });
-
-    }
-
-    /**
      * Update all node positions and edge paths.
      *
      * @param eventSource the event source used for render events (default: `EventSource.API`)
@@ -2623,12 +1756,12 @@ export default class GraphEditor extends HTMLElement {
         this.edgesGroup
             .selectAll('g.edge-group:not(.dragged)')
             .data(this._edges, edgeId)
-            .call(this.updateEdgePositions.bind(this));
+            .call(this.edgeRenderer.updateEdgePositions.bind(this.edgeRenderer));
 
         this.edgesGroup
             .selectAll('g.edge-group.dragged')
             .data(this.draggedEdges, edgeId)
-            .call(this.updateEdgePositions.bind(this));
+            .call(this.edgeRenderer.updateEdgePositions.bind(this.edgeRenderer));
 
         this.onRender(eventSource, 'positions');
     }
@@ -2641,249 +1774,6 @@ export default class GraphEditor extends HTMLElement {
     public updateHighlights(): void {
         this.updateNodeHighligts();
         this.updateEdgeHighligts();
-    }
-
-    /**
-     * Create a new dragged edge from a source node.
-     *
-     * @param sourceNode node that edge was dragged from
-     */
-    public createDraggedEdge(event: Event, sourceNode: Node): DraggedEdge { // FIXME visibility
-        const validTargets = new Set<string>();
-        this._nodes.forEach(node => validTargets.add(node.id.toString()));
-        this.objectCache.getEdgesBySource(sourceNode.id).forEach(edge => validTargets.delete(edge.target.toString()));
-        validTargets.delete(sourceNode.id.toString());
-        let draggedEdge: DraggedEdge = {
-            id: sourceNode.id.toString() + Date.now().toString(),
-            source: sourceNode.id,
-            target: null,
-            validTargets: validTargets,
-            currentTarget: { x: (event as any).x, y: (event as any).y },
-        };
-        if (this.onCreateDraggedEdge != null) {
-            draggedEdge = this.onCreateDraggedEdge(draggedEdge);
-            if (draggedEdge == null) {
-                return null;
-            }
-        }
-        this.draggedEdges.push(draggedEdge);
-        return draggedEdge;
-    }
-
-    /**
-     * Create a dragged edge from an existing edge.
-     *
-     * @param edge existing edge
-     * @param reverseEdgeDirection reverse the direction of the returned edge
-     */
-    // eslint-disable-next-line complexity
-    private createDraggedEdgeFromExistingEdge(event: Event, edge: Edge, reverseEdgeDirection: boolean = false): DraggedEdge {
-        const validTargets = new Set<string>();
-        this._nodes.forEach(node => validTargets.add(node.id.toString()));
-        const source = reverseEdgeDirection ? edge.target : edge.source;
-        this.objectCache.getEdgesBySource(source).forEach(edgeOutgoing => {
-            if (edgeId(edge) !== edgeId(edgeOutgoing)) {
-                validTargets.delete(edgeOutgoing.target.toString());
-            }
-        });
-        let draggedEdge: DraggedEdge = {
-            id: source.toString() + Date.now().toString(),
-            createdFrom: edgeId(edge),
-            source: reverseEdgeDirection ? edge.target : edge.source,
-            target: null,
-            type: edge.type,
-            pathType: edge.pathType,
-            validTargets: validTargets,
-            currentTarget: { x: (event as any).x, y: (event as any).y },
-            markers: [],
-            texts: [],
-            dragHandles: null,
-            isBidirectional: edge.isBidirectional,
-        };
-        for (const key in edge) {
-            if (edge.hasOwnProperty(key) && key !== 'id' && key !== 'type' && key !== 'createdFrom' && key !== 'sourceHandle' &&
-                key !== 'targetHandle' && key !== 'source' && key !== 'target' && key !== 'validTargets' && key !== 'currentTarget') {
-                try {
-                    draggedEdge[key] = JSON.parse(JSON.stringify(edge[key]));
-                } catch (error) {
-                    draggedEdge[key] = edge.key;
-                }
-            }
-        }
-        if (reverseEdgeDirection) {
-            const flipDirections = (element: PathPositionRotationAndScale) => {
-                // flip positionOnLine
-                if (element.positionOnLine === 'start') {
-                    element.positionOnLine = 'end';
-                } else if (element.positionOnLine === 'end') {
-                    element.positionOnLine = 'start';
-                } else {
-                    element.positionOnLine = 1 - element.positionOnLine;
-                }
-                // rotate relative rotation by 180°
-                if (element.relativeRotation != null) {
-                    element.relativeRotation += element.relativeRotation > 180 ? -180 : 180;
-                }
-                // flip absolutePositionOnLine
-                if (element.absolutePositionOnLine != null) {
-                    element.absolutePositionOnLine = -element.absolutePositionOnLine;
-                }
-            };
-            draggedEdge.markers.forEach(flipDirections);
-            draggedEdge.texts.forEach(flipDirections);
-            draggedEdge.dragHandles.forEach(handle => {
-                flipDirections(handle);
-                // flip isReverseHandle
-                handle.isReverseHandle = !handle.isReverseHandle;
-            });
-            // flip end markers
-            const temp = draggedEdge.markerEnd;
-            if (draggedEdge.markerStart != null) {
-                draggedEdge.markerEnd = draggedEdge.markerStart;
-            }
-            if (temp != null) {
-                draggedEdge.markerStart = temp;
-            }
-        }
-        if (this.onCreateDraggedEdge != null) {
-            draggedEdge = this.onCreateDraggedEdge(draggedEdge);
-            if (draggedEdge == null) {
-                return null;
-            }
-        }
-        this.draggedEdges.push(draggedEdge);
-        return draggedEdge;
-    }
-
-    /**
-     * Update dragged edge on drag event.
-     */
-    // eslint-disable-next-line complexity
-    public updateDraggedEdge(event: Event, edge: DraggedEdge, capturingGroup?: string) { //FIXME visibility
-        const oldTarget = edge.target;
-        edge.target = null;
-        edge.currentTarget.x = (event as any).x;
-        edge.currentTarget.y = (event as any).y;
-
-        const sourceEvent = (event as any).sourceEvent;
-        const possibleTargetNodes = this.getNodesFromPoint(sourceEvent.clientX, sourceEvent.clientY);
-        if (possibleTargetNodes.length > 0) {
-            const targetNode = possibleTargetNodes[0];
-            const targetNodeId = targetNode.id.toString();
-
-            // validate target
-            let isValidTarget = true;
-
-            // allow target to be source
-            // isValidTarget = isValidTarget && (edge.source.toString() === targetNodeId);
-
-            // check group capture
-            const targetGroupCapturingEdge = this.groupingManager.getGroupCapturingIncomingEdge(targetNode);
-
-            if (targetGroupCapturingEdge == null) {
-                // no group capture, node must be a valid target
-                isValidTarget = isValidTarget && edge.validTargets.has(targetNodeId);
-            } else {
-                // group capture, group must be a valid target
-                isValidTarget = isValidTarget && edge.validTargets.has(targetGroupCapturingEdge);
-            }
-
-            if (isValidTarget) {
-                // initially always set the target node
-                edge.target = targetNodeId;
-            } else {
-                // remove target if no valid target found
-                edge.target = null;
-            }
-
-            // handle group captures
-            if (isValidTarget && targetGroupCapturingEdge != null) {
-                const targetGroupBehaviour = this.groupingManager.getGroupBehaviourOf(targetGroupCapturingEdge);
-                const targetGroupNode = this.getNode(targetGroupCapturingEdge);
-                if (targetGroupBehaviour?.delegateIncomingEdgeTargetToNode != null && targetGroupNode != null) {
-                    const newTarget = targetGroupBehaviour.delegateIncomingEdgeTargetToNode(targetGroupCapturingEdge, targetGroupNode, edge, this);
-                    if (newTarget != null && newTarget !== '' && this.getNode(newTarget) !== null) {
-                        edge.target = newTarget;
-                    }
-                }
-            }
-        }
-
-        // dispatch event if target changed and handle source group link capture
-        if (edge.target !== oldTarget) {
-            if (capturingGroup != null) {
-                const groupBehaviour = this.groupingManager.getGroupBehaviourOf(capturingGroup);
-                const groupNode = this.getNode(capturingGroup);
-                if (groupBehaviour != null && groupNode != null && groupBehaviour.delegateOutgoingEdgeSourceToNode != null) {
-                    const newSource = groupBehaviour.delegateOutgoingEdgeSourceToNode(capturingGroup, groupNode, edge, this);
-                    if (newSource != null && newSource !== '' && this.getNode(newSource) !== null) {
-                        edge.source = newSource;
-                    }
-                }
-            }
-            if (this.onDraggedEdgeTargetChange != null) {
-                const source = this.objectCache.getNode(edge.source);
-                const target = edge.target != null ? this.objectCache.getNode(edge.target) : null;
-                this.onDraggedEdgeTargetChange(edge, source, target);
-            }
-        }
-    }
-
-    /**
-     * Drop dragged edge.
-     *
-     * @param edge the edge that was dropped
-     * @param isReversedEdge true if the edge is bidirectional and was dragged
-     *      from a reverse handle flipping the edge direction
-     */
-    // eslint-disable-next-line complexity
-    public dropDraggedEdge(event: Event, edge: DraggedEdge, isReversedEdge: boolean) { // FIXME visibility
-        let updateEdgeCache = false;
-        const existingEdge = this.objectCache.getEdge(edge.createdFrom);
-        let existingTarget = existingEdge?.target.toString();
-        if (isReversedEdge) {
-            existingTarget = existingEdge?.source.toString();
-        }
-        if (edge.createdFrom != null) {
-            if (edge.target?.toString() !== existingTarget) {
-                // only remove original edge if target of dropped edge is different then original target
-                const i = this._edges.findIndex(e => edgeId(e) === edge.createdFrom);
-                if (this.onEdgeRemove(this._edges[i], EventSource.USER_INTERACTION)) {
-                    this._edges.splice(i, 1);
-                    updateEdgeCache = true;
-                }
-            }
-        }
-
-        const index = this.draggedEdges.findIndex(e => e.id === edge.id);
-        this.draggedEdges.splice(index, 1);
-        this.updateDraggedEdgeGroups();
-        if (edge.target != null) {
-            // dragged edge has a target
-            let finalEdge: Edge = edge;
-            delete finalEdge.id;
-            if (this.onDropDraggedEdge != null) {
-                finalEdge = this.onDropDraggedEdge(edge, this.objectCache.getNode(edge.source),
-                    this.objectCache.getNode(edge.target));
-            }
-            if (edge.createdFrom != null && edge.target === existingTarget) {
-                // edge was dropped on the node that was the original target for the edge
-                this.completeRender(false, EventSource.USER_INTERACTION);
-            } else {
-                if (this.onEdgeCreate(edge, EventSource.USER_INTERACTION)) {
-                    this._edges.push(edge);
-                    updateEdgeCache = true;
-                }
-            }
-        } else {
-            this.onEdgeDrop(edge, { x: (event as any).x, y: (event as any).y });
-        }
-        if (updateEdgeCache) {
-            this.objectCache.updateEdgeCache(this._edges);
-            this.completeRender(false, EventSource.USER_INTERACTION);
-        } else {
-            this.updateEdgeGroupClasses();
-        }
     }
 
     /**
@@ -2907,35 +1797,6 @@ export default class GraphEditor extends HTMLElement {
         return this.dispatchEvent(ev);
     }
 
-
-    /**
-     * Callback for creating edgedrop events.
-     *
-     * The event is only for dragged edges that are dropped in the void.
-     *
-     * @param edge the dropped dragged edge
-     * @param dropPosition the position where the edge was dropped at
-     * @returns false if event was cancelled
-     */
-    private onEdgeDrop(edge: DraggedEdge, dropPosition: Point) {
-        const detail: any = {
-            eventSource: EventSource.USER_INTERACTION,
-            edge: edge,
-            sourceNode: this.objectCache.getNode(edge.source),
-            dropPosition: dropPosition,
-        };
-        if (edge.createdFrom != null && edge.createdFrom !== '') {
-            detail.originalEdge = this.objectCache.getEdge(edge.createdFrom);
-        }
-        const ev = new CustomEvent('edgedrop', {
-            bubbles: true,
-            composed: true,
-            cancelable: false,
-            detail: detail,
-        });
-        return this.dispatchEvent(ev);
-    }
-
     /**
      * Callback for creating edgeremove events.
      *
@@ -2955,83 +1816,6 @@ export default class GraphEditor extends HTMLElement {
             },
         });
         return this.dispatchEvent(ev);
-    }
-
-    /**
-     * Callback on edges for click event.
-     *
-     * @param edgeDatum Corresponding datum of edge
-     */
-    private onEdgeClick(event: Event, edgeDatum) {
-        const eventDetail: any = {};
-        eventDetail.eventSource = EventSource.USER_INTERACTION;
-        const path = event.composedPath();
-        if (path != null) {
-            let i = 0;
-            let target;
-            // search in event path for data-click attribute
-            while (i === 0 || target != null && i < path.length) {
-                target = select(path[i] as any);
-                const key = target.attr('data-click');
-                if (key != null) {
-                    eventDetail.key = key;
-                    break;
-                }
-                if (target.classed('edge-group')) {
-                    // reached edge boundary in dom
-                    break;
-                }
-                i++;
-            }
-        }
-        eventDetail.sourceEvent = event;
-        eventDetail.edge = edgeDatum;
-        const ev = new CustomEvent('edgeclick', { bubbles: true, composed: true, cancelable: true, detail: eventDetail });
-        if (!this.dispatchEvent(ev)) {
-            return; // prevent default / event cancelled
-        }
-    }
-
-    /**
-     * Create and dispatch 'edgetextdragstart' and 'edgetextdragend' events.
-     *
-     * @param eventType the type of the event
-     * @param textComponent The text component that was dragged.
-     * @param edge The edge the text component belongs to.
-     * @param eventSource the event source
-     */
-    private onEdgeTextDrag(eventType: 'start' | 'end', textComponent: TextComponent, edge: Edge, eventSource) {
-        const ev = new CustomEvent(`edgetextdrag${eventType}`, {
-            bubbles: true,
-            composed: true,
-            cancelable: false,
-            detail: {
-                eventSource: eventSource,
-                text: textComponent,
-                edge: edge,
-            },
-        });
-        this.dispatchEvent(ev);
-    }
-
-    /**
-     * Callback for creating edgetextpositionchange events.
-     *
-     * @param textComponent The text component that was dragged.
-     * @param edge The edge the text component belongs to.
-     */
-    private onEdgeTextPositionChange(textComponent: TextComponent, edge: Edge) {
-        const ev = new CustomEvent('edgetextpositionchange', {
-            bubbles: true,
-            composed: true,
-            cancelable: false,
-            detail: {
-                eventSource: EventSource.USER_INTERACTION,
-                text: textComponent,
-                edge: edge,
-            },
-        });
-        this.dispatchEvent(ev);
     }
 
     /**
@@ -3266,27 +2050,11 @@ export default class GraphEditor extends HTMLElement {
     /**
      * Calculate highlighted edges and update their classes.
      */
-    private updateEdgeHighligts(edgeSelection?: Selection<SVGGElement, Edge, any, unknown>) {
+    public updateEdgeHighligts(edgeSelection?: Selection<SVGGElement, Edge, any, unknown>) {
         if (edgeSelection == null) {
-            edgeSelection = this.edgesGroup
-                .selectAll<SVGGElement, Edge>('g.edge-group:not(.dragged)')
-                .data<Edge>(this._edges, edgeId);
+            edgeSelection = this.getEdgeSelection();
         }
-
-        let nodes: Set<number | string> = new Set();
-
-        if (this._selectedLinkSource != null) {
-            nodes.add(this._selectedLinkSource);
-            if (this._selectedLinkTarget != null) {
-                nodes.add(this._selectedLinkTarget);
-            }
-        } else {
-            nodes = this.hovered;
-        }
-
-        edgeSelection
-            .classed('highlight-outgoing', (d) => nodes.has(d.source))
-            .classed('highlight-incoming', (d) => nodes.has(d.target));
+        this.edgeRenderer.updateEdgeHighligts(edgeSelection, this.hovered, this._selectedLinkSource, this._selectedLinkTarget);
     }
 
     /**
